Move internal quoted.Type implementations to internal package
