+++ compiler/src/dotty/tools/dotc/core/tasty/TreeUnpickler.scala
@@ -33,8 +33,7 @@ import config.Printers.pickling
 import core.quoted.PickledQuotes
 
 import scala.quoted
-import scala.quoted.Types.TreeType
-import scala.internal.quoted.TastyTreeExpr
+import scala.internal.quoted.{TastyTreeExpr, TreeType}
 import scala.annotation.constructorOnly
 import scala.annotation.internal.sharable
